fix: getPathToRootOrCompaction
